LLAMA-5505 - Fix loss of ReadyForMoreSamples notifications
Reason for change: Notification messages lost
Test Procedure: Check videos play without issues and no AV freeze is
seen
Risks: Low

When delegating the notification of "ReadyForMoreSamples" to another
thread via
the "MainThreadNotifier" class, if a prior notification request is still
pending, the new request is dropped. As notification requests can occur
concurrently when multiple streams are present (e.g. audio and video), a
notification may get lost, because the notify only tracks pending
notifications per notification type, and does not have knowledge of
streams.
I.e. if the notifier has a pending "ReadyForMoreSamples" notification
for stream 1 and we are requesting a notification for stream 2, then the
notification for stream 2 is ignored/dropped.

While this kind of notification may be triggered again, there are corner
cases when such re-trigger is not happening if it was already requested
(e.g. empty queue). To ensure notifications are triggered for each
stream, we use stream type specific notifications.

An observed effect of this issue is video freezing with audio continuing

Signed-off-by: Filipe Norte <filipe.norte@sky.uk>
Change-Id: Ica5d0691fefe9602357d1d26429483b0c618f833