RDKTV-15263: Audio heard during trickplay when VG is ON
Reason for change: appletv+ mutes the player but it is still in playing
state
When VG plays audio ducking happens (appletv+ is SpeechSynthesis)
amlogic does not check mute before setting the volume. So the volume
setting that happens in (2) overrides the mute.
setPageMediaVolume()" is required only in BCOM platforms
Test Procedure: Enable VG
1. Test apple tv+ and other webapps during trickplay
2. Test appletv+ and other webapps for audio ducking feature
Risks: NA

Signed-off-by: smathe135 <Simi_Mathew@comcast.com>
Change-Id: I66ec100583f0b4fcf58a687b8c18c04a9aa08a5d