DELIA-58021: WPE Packager fails when LXC XRE is enabled
Reason for change: Packager library is trying to write the
rdm_status.log file. While running inside the container, it is not able
to open this log file in write mode as the container is non root and the
syslog-ng has created this log file with 0644 permission mode. As it is
not able to write the log file, packager process is getting crashed.
Hence modifying the permission of output files as 0664 so that group
users can also write on the log file.
Test Procedure: No packager crash should be observed.
Risks: Low

Signed-off-by: Arjun <arjun_daasuramdass@comcast.com>
(cherry picked from commit 6fc0eede2da6525dd18fdb4135d6b27ba0381d24)
Change-Id: I76fad7a480f92e7172df387dbece0830df3c555e