DELIA-58562 : iHeartRadio app assets stuck in loading icon
Reason for change: The following commit is being reverted
 https://github.com/WebPlatformForEmbedded/WPEWebKit/commit/282000d7b705f554ce2c124f9bd0bd236db1676d

as a workaround for the issue "Spinning wheel seen on iheart in realtek
devices"
The change
        (m_bufferingPercentage == 0 && (m_maxTimeLoaded -
        playbackPosition() < MediaTime::createWithDouble(2,
        GST_SECOND))))
is causing a constant spinning wheel in iheart assets(live) as
the m_maxTimeLoaded is never updated and queue2 element very
frequently sends a buffering message because of the waythe
rtkaudiosink operates. This causes the readystate to flipflop
between HaveCurrentData and HaveEnoughData and hence the
spinning wheel.
Test Procedure: Playback iHeartRadio assets to see if spinning
wheel issues is seen. Regression testing for apps using progressive
playback - pandora, iheart, xite, adventure2learning, France24
Risks: Low
Priority: P1
Signed-off-by: Simi Mathew <simim@tataelxsi.co.in>

Change-Id: Idaf52a752692ae3bb28870f521c255fcd53b5d11